for Stefano, forever ago
